prevent page breaks in awkward locations
The API for long sections is slightly unfortunate: the body must be
given as a list instead of a do block containing a for loop, so that we
can treat the first item specially. Alas, attaching to the header the
CSS "break-after: avoid;" doesn't seem to work, so we instead have to
use a "break-inside: avoid;" region which includes both the title and
the first element.